requirements.txt
pydantic-ai[openai]>=0.0.20
gradio>=4.0.0
python-dotenv>=1.0.0
ddgs>=9.0.0
